% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.genotypes.R
\name{decode.genotypes}
\alias{decode.genotypes}
\title{Creates contrasts that relate to a single locus}
\usage{
decode.genotypes(genotype)
}
\arguments{
\item{genotype}{The vector \code{genotype} holds two-letter codes for the three genotypes. For example, the values may be AA, Aa and aa.}
}
\value{
\itemize{
    \item{oset}{The offset values that would be appropriate, in the multiplicative version of the model, if there was just this one locus.}
    \item{ma}{A contrast for the Hardy-Weinberg model, at this locus.}
    \item{maa}{A contrast that measures departure from the Hardy-Weinberg model, at this locus.}
    \item{types}{A vector of length three whose elements are the two-letter codes used for the three genotypes.}
}
}
\description{
Vectors ma, maa and oset (offset) are returned.
}
\note{
Called by \code{make.contrasts}
}
\examples{
decode.genotypes(rep(c("AA","Aa","aa"),2))

}
\seealso{
\code{hwde}
}
\author{
J.H. Maindonald
}
