% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwexact.R
\name{hwexact}
\alias{hwexact}
\title{Exact test of HWE}
\usage{
hwexact(
  obs.hom1 = 68,
  obs.hets = 28,
  obs.hom2 = 4,
  data = NULL,
  loci = "locus1",
  observed = "Observed"
)
}
\arguments{
\item{obs.hom1}{Observed number of individuals homozygous for the common allele.}

\item{obs.hets}{Observed number of heterozygous individuals.}

\item{obs.hom2}{Observed number of individuals homozygous for the rare allele.}

\item{data}{If not \code{NULL}, specifies a data frame holding the data.}

\item{loci}{If \code{data} is not \code{NULL}, column that holds some combination of symbols of the form \code{AA}, \code{Aa} and \code{aa}.}

\item{observed}{Column that holds the frequencies, if \code{data} is not \code{NULL}.}
}
\value{
The exact p-value testing departure from Hardy-Weinberg Equilibrium, conditional on the observed relative numbers of alleles of the two types.
}
\description{
This function implements an exact SNP test of Hardy-Weinberg equilibrium.
}
\examples{
hwexact(68, 28, 4)

data(IndianIrish)
hwexact(data = IndianIrish, loci = 'locus1', observed = 'Observed')

}
\references{
Wigginton, J.E.; Cutler, D.J. & Abecasis, G.R.; A Note on exact tests of Hardy-Weinberg equilibrium; American Journal of Human
}
\author{
Randall Johnson -- adapted from code by Wigginton et al Genetics, 2005, 76, 887-93
}
