% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IndianIrish}
\alias{IndianIrish}
\title{Observed genotype frequencies at MN and S loci, for 2 populations}
\format{
This data frame contains the following columns:
\describe{
  \item{Population}{Factor with levels: \code{Indian} and \code{Irish} }
  \item{locus1}{Factor with levels: \code{MM}, \code{MN} and \code{NN}}
  \item{locus2}{Factor with levels: \code{SS}, \code{Ss} and \code{ss}}
  \item{Observed}{a numeric vector giving the frequency for each category of the tale}
}
}
\source{
Mourant et al (1977) and Huttley and Wilson (2000).
}
\usage{
IndianIrish
}
\description{
The \code{IndianIrish} data frame has 18 rows and 4 columns. The data are genotype frequencies for two locations, for Xavante Indian and Irish populations respectively.
}
\examples{
data(IndianIrish)
hwde(data=IndianIrish)
}
\references{
\itemize{
  \item Huttley, G.A. and Wilson, S.R. 2000. Testing for concordant equilibria between population samples. \emph{Genetics} \bold{156}, 2127-2135.
  \item Mourant, A.E., Kopec, A.C. and Domaniewska-Sobczak, K. 1976. \emph{The Distribution of the Human Blood Groups and Other Polymorphisms.} Oxford University Press.
  \item Weir, B.S. 1996.  \emph{Genetic Data Analysis II.} Sinauer.
}
}
\seealso{
\code{\link{hwde}}
}
\keyword{datasets}
