% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.contrasts.R
\name{make.contrasts}
\alias{make.contrasts}
\title{Calculate contrasts required by the function}
\usage{
make.contrasts(data = data[, loci], allele.chars = letters)
}
\arguments{
\item{data}{Data frame, with one column per locus, and as many rows as there are observations. Genotypes are represented using a two-letter code, e.g., AA, Aa, aa}

\item{allele.chars}{By default the letters "a", "b", ..., one for each locus, are used.}
}
\value{
\itemize{
  \item{contrasts.df}{Data frame whose columns hold the contrasts that will be explanatory variables for the Poisson regression model.}
  \item{oset}{Offset vector for the GLM model.}
  \item{list.columns}{List in which each element is a text string that holds the codes for a sequence of model terms.}
}
}
\description{
Calculates the contrasts that \code{hwde} uses in fitting models for genotypic disequilibrium. At present these are limited to first order interaction terms between the distinct loci, possibly with distinct values foe each different level of a grouping factor.
}
\details{
Any pair of letters can be used for the two alleles. The setting of \code{allele.chars} determines the coding used in the model formulae. This function is called by \code{hwde}
}
\examples{
make.contrasts(data=c("AA","Aa","aa"))

}
\seealso{
\code{identify.genotypes}, called by \code{make.contrasts}
}
\author{
J.H. Maindonald
}
